server/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
duckdb==1.1.0
python-jose==3.3.0
pydantic==2.8.2
